requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
imgkit==1.2.2
Markdown==3.3.4
Pillow==8.2.0
python-frontmatter==1.0.0
PyYAML==5.4.1
six==1.16.0
soupsieve==2.2.1
